requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Transcription
openai-whisper>=20231117
torch>=2.0.0

# PHI Redaction
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# LLM
requests>=2.31.0

# Database
# (sqlite3 is built into Python)

# PDF
fpdf2>=2.7.0

# Demo UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
httpx>=0.25.0
